refacto(archi): move Routing, Home, and Http/Cache
