+++ src/Auth/Domain/Authenticator/LoginFormAuthenticator.php
@@ -2,7 +2,7 @@
 
 namespace App\Auth\Domain\Authenticator;
 
-use App\Domain\Routing\ValueObject\RouteName;
+use App\Routing\Domain\ValueObject\RouteName;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
