+++ src/Shared/Http/Cache/HttpCache.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace App\Http\Cache;
+namespace App\Shared\Http\Cache;
 
-use App\Http\Cache\Adapter\HttpCacheAdapter;
-use App\Http\Cache\ValueObject\CacheItem;
+use App\Shared\Http\Cache\Adapter\HttpCacheAdapter;
+use App\Shared\Http\Cache\ValueObject\CacheItem;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
