+++ templates/blog/base.html.twig
@@ -6,7 +6,7 @@
         '@context': "https://schema.org",
         '@type': "Blog",
         name: "Hugo Alliaume's Blog",
-        url: url(constant('App\\Domain\\Routing\\ValueObject\\RouteName::BLOG_HOME')),
+        url: url(constant('App\\Routing\\Domain\\ValueObject\\RouteName::BLOG_HOME')),
     }) }}
 {% endblock %}
