+++ templates/blog/home.html.twig
@@ -19,7 +19,7 @@
                 <article class="py-4 sm:py-6 first:pt-0 first:sm:pt-0 last:pb-0 last:pb-0">
                     <header>
                         <h2 class="text-2xl sm:text-3xl font-bold mb-4">
-                            <a href="{{ path(constant('App\\Domain\\Routing\\ValueObject\\RouteName::BLOG_POST_VIEW'), {slug: post.slug}) }}">
+                            <a href="{{ path(constant('App\\Routing\\Domain\\ValueObject\\RouteName::BLOG_POST_VIEW'), {slug: post.slug}) }}">
                                 {{ post.title }}
                             </a>
                         </h2>
