+++ templates/blog/tags/view/index.html.twig
@@ -15,7 +15,7 @@
                     <article class="py-4 sm:py-6 first:pt-0 first:sm:pt-0 last:pb-0 last:pb-0">
                         <header>
                             <h2 class="text-2xl sm:text-3xl font-bold mb-4">
-                                <a href="{{ prefetch(path(constant('App\\Domain\\Routing\\ValueObject\\RouteName::BLOG_POST_VIEW'), {slug: post.slug})) }}">
+                                <a href="{{ prefetch(path(constant('App\\Routing\\Domain\\ValueObject\\RouteName::BLOG_POST_VIEW'), {slug: post.slug})) }}">
                                     {{ post.title }}
                                 </a>
                             </h2>
