+++ templates/components/CloudOfTags.html.twig
@@ -5,7 +5,7 @@
         <li>
             <twig:Badge>
                 <a
-                    href="{{ path(constant('App\\Domain\\Routing\\ValueObject\\RouteName::BLOG_TAG_VIEW'), {tag: tag.tag}) }}"
+                    href="{{ path(constant('App\\Routing\\Domain\\ValueObject\\RouteName::BLOG_TAG_VIEW'), {tag: tag.tag}) }}"
                     title="{{ ('View all posts tagged with "' ~ tag.tag ~ '"')|e('html_attr') }}"
                 >
                     {{- tag.tag ~ ' (' ~ tag.occurrences ~ ')' -}}
