+++ templates/components/PostMeta.html.twig
@@ -11,7 +11,7 @@
         <ul class="flex flex-row flex-wrap gap-1">
             {% for tag in post.tags %}
                 <li>
-                    <twig:Badge size="sm" href="{{ path(constant('App\\Domain\\Routing\\ValueObject\\RouteName::BLOG_TAG_VIEW'), {tag}) }}">
+                    <twig:Badge size="sm" href="{{ path(constant('App\\Routing\\Domain\\ValueObject\\RouteName::BLOG_TAG_VIEW'), {tag}) }}">
                         {{- tag -}}
                     </twig:Badge>
                 </li>
